Add the synthetic data generator. It works for the annotation specification example (eg1,eg12,eg13,eg2).
However, one thing need to be improved.
For observatioins with identifying context chain 
The distinct factor now just works as a lower bound of the distinct number of observations. 
I.e., if the distinct factor is 0.5, and n=100
for observation types without identifying context.
It can guarantee that the distinct number of keys are 50. 

However, for observation types with identifying context (chain), 
the current program can only guarantee 
the generated number of distinct keys >=50
This problem is caused by the dependency relationship between the measurement keys. I need to rethink about a way to generate this part to guarantee the distinct factor.