1. Entity: added some substances (compounds) CompoundSubstance   Ammonium
2. and Measurements (concentration classes) measurement for the Ammonium entity
concentration
 ConcentrationOfAmmonium
   SBC_concentrationOfAmmonium (2) types.


These are intended to be similar to the measurements defined in the SBC oboe-beta, but per recent discussions. 

2. The kelpPercentNitrogen now passes the Pellet reasoner

There is only one set of substance-measurement, so far, to see how they work out. Looking forward to discussing! particularly, how this might work
"entity=Ammonium, context=SalineWater, characteristic=AmountOfSubstanceConcentration"

instead of just
"entity=water, characteristic=AmountOfSubstanceConcentration"
and Ammonium is buried in the Protocol.